Implement `chargeURI` with `TokenUcrx`
