--- src/impl/uri-chargeable.placement.ts
@@ -1,7 +1,0 @@
-import { URIChargeable } from '../schema/uri-chargeable.js';
-
-export const ANY_CHARGE_PLACEMENT: URIChargeable.Any = {
-  as: undefined,
-  omitCommaBefore: undefined,
-  omitCommaAfter: undefined,
-};
