+++ src/rx/mod.ts
@@ -1,3 +1,5 @@
+export * from './charge-uri.js';
+export * from './churi.tag.js';
 export * from './entity.ucrx-handler.js';
 export * from './entity.ucrx.js';
 export * from './opaque.ucrx.js';
