+++ src/schema/mod.ts
@@ -1,6 +1,4 @@
 export * from './boolean/mod.js';
-export * from './charge-uri.js';
-export * from './churi.tag.js';
 export * from './entity/mod.js';
 export * from './list/mod.js';
 export * from './map/mod.js';
@@ -16,4 +14,3 @@ export * from './uc-schema.js';
 export * from './uc-serializer.js';
 export * from './unknown/mod.js';
 export * from './uri-charge/mod.js';
-export * from './uri-chargeable.js';
