+++ src/schema/uri-charge/uri-charge.ts
@@ -1,7 +1,6 @@
 import { Ucrx } from '../../rx/ucrx.js';
 import { Uctx } from '../../rx/uctx.js';
 import { UcUnknown } from '../unknown/uc-unknown.js';
-import { URIChargeable } from '../uri-chargeable.js';
 
 /**
  * Generic URI charge representation.
@@ -10,7 +9,7 @@ import { URIChargeable } from '../uri-chargeable.js';
  * and {@link URICharge.List lists}. When part of the charge is not available, corresponding property or method returns
  * {@link URICharge.none}, which means the absence of charge.
  */
-export abstract class URICharge implements URIChargeable, Uctx {
+export abstract class URICharge implements Uctx {
 
   /**
    * URI charge instance representing the absent charge.
@@ -142,15 +141,6 @@ export abstract class URICharge implements URIChargeable, Uctx {
   abstract keys(): IterableIterator<string>;
 
   /**
-   * Encodes this URI charge.
-   *
-   * @param placement -  The supposed placement of encoded value.
-   *
-   * @returns String with encoded value, or `undefined` for {@link URICharge.none absent charge}.
-   */
-  abstract chargeURI(placement: URIChargeable.Placement): string | undefined;
-
-  /**
    * Transfers this URI charge to the given receiver.
    *
    * @param rx - Charge receiver.
@@ -213,18 +203,14 @@ export namespace URICharge {
     get(key: string): None;
     entries(): IterableIterator<never>;
     keys(): IterableIterator<never>;
-
-    chargeURI(_placement: URIChargeable.Placement): undefined;
   }
 
   /**
    * URI charge that represents {@link URICharge#isSome something}, in contrast to {@link URICharge.None none}.
    */
-  export interface Some extends URICharge, URIChargeable {
+  export interface Some extends URICharge {
     isNone(): false;
     isSome(): true;
-
-    chargeURI(placement: URIChargeable.Placement): string;
   }
 
   /**
@@ -251,8 +237,6 @@ export namespace URICharge {
     get(key: string): None;
     entries(): IterableIterator<never>;
     keys(): IterableIterator<never>;
-
-    chargeURI(placement: URIChargeable.Placement): string;
   }
 
   /**
@@ -267,8 +251,6 @@ export namespace URICharge {
     get(key: string): None;
     entries(): IterableIterator<never>;
     keys(): IterableIterator<never>;
-
-    chargeURI(placement: URIChargeable.Placement): string;
   }
 
   /**
@@ -291,8 +273,6 @@ export namespace URICharge {
     at(index: 0 | -1): this;
     at(index: number): this | None;
     list(): IterableIterator<this>;
-
-    chargeURI(placement: URIChargeable.Placement): string;
   }
 }
 
@@ -352,10 +332,6 @@ class URICharge$None extends URICharge implements URICharge.None {
 
   *keys(): IterableIterator<never> {
     // Not a map.
-  }
-
-  override chargeURI(_placement: URIChargeable.Placement): undefined {
-    return;
   }
 
   override toUc(_rx: Ucrx): void {
